+++ src/main/java/io/company/brewcraft/repository/FinishedGoodLotRepository.java
@@ -17,5 +17,5 @@ public interface FinishedGoodLotRepository extends JpaRepository<FinishedGoodLot
     @Override
     @Modifying
     @Query("delete from finished_good_lot i where i.id in (:ids)")
-    int deleteByIds(Iterable<Long> ids);
+    long deleteByIds(Iterable<Long> ids);
 }
