+++ src/main/java/io/company/brewcraft/repository/ShipmentRepository.java
@@ -17,5 +17,5 @@ public interface ShipmentRepository extends JpaRepository<Shipment, Long>, JpaSp
     @Override
     @Modifying
     @Query("delete from shipment s where s.id in (:ids)")
-    int deleteByIds(Iterable<Long> ids);
+    long deleteByIds(Iterable<Long> ids);
 }
