+++ src/main/java/io/company/brewcraft/service/FinishedGoodLotService.java
@@ -109,12 +109,12 @@ public class FinishedGoodLotService extends BaseService implements CrudService<L
     }
 
     @Override
-    public int delete(Set<Long> ids) {
+    public long delete(Set<Long> ids) {
         return this.repoService.delete(ids);
     }
 
     @Override
-    public int delete(Long id) {
+    public long delete(Long id) {
         return this.repoService.delete(id);
     }
