+++ src/main/java/io/company/brewcraft/service/IaasObjectStoreService.java
@@ -36,14 +36,14 @@ public class IaasObjectStoreService extends BaseService implements CrudService<S
     }
 
     @Override
-    public int delete(Set<String> ids) {
+    public long delete(Set<String> ids) {
         this.iaasRepo.delete(ids);
 
         return ids.size();
     }
 
     @Override
-    public int delete(String id) {
+    public long delete(String id) {
         this.iaasRepo.delete(List.of(id));
 
         return 1;
