+++ src/test/java/io/company/brewcraft/controller/ProcurementControllerTest.java
@@ -186,7 +186,7 @@ public class ProcurementControllerTest {
         ArgumentCaptor<Set<ProcurementId>> captor = ArgumentCaptor.forClass(Set.class);
         doReturn(3).when(mCrudController).delete(captor.capture());
 
-        int count = controller.delete(Set.of(new ProcurementIdDto(1L, 1L), new ProcurementIdDto(2L, 2L), new ProcurementIdDto(3L, 3L)));
+        long count = controller.delete(Set.of(new ProcurementIdDto(1L, 1L), new ProcurementIdDto(2L, 2L), new ProcurementIdDto(3L, 3L)));
 
         assertEquals(3, count);
         assertThat(captor.getValue()).hasSameElementsAs(Set.of(new ProcurementId(1L, 1L), new ProcurementId(2L, 2L), new ProcurementId(3L, 3L)));
